Add common exception handling, return concrete types as preference

+++ UKSF.Api.Integrations.Teamspeak/Models/TeamspeakClient.cs
@@ -1,8 +1,10 @@
-﻿namespace UKSF.Api.Teamspeak.Models {
-    public class TeamspeakClient {
-        public double ChannelId;
+﻿namespace UKSF.Api.Teamspeak.Models
+{
+    public class TeamspeakClient
+    {
+        public int ChannelId;
         public string ChannelName;
-        public double ClientDbId;
+        public int ClientDbId;
         public string ClientName;
     }
 }
